pyproject.toml
[tool.poetry]
name = "facebookfriendnetworkanalyzer"
version = "0.1.0"
description = ""
authors = ["Rubén Chuliá Mena <rubchume@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
selenium = "^3.141.0"
numpy = "^1.21.0"
matplotlib = "^3.4.2"
pandas = "1.2.5"
networkx = "^2.5.1"
Django = "^3.2.5"
django-background-tasks = "^1.2.5"
pytest-django = "^4.4.0"
plotly = "^5.1.0"
webdriver-manager = "^3.4.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
nox = "^2021.6.12"
safety = "^1.10.3"
flake8 = "^3.9.2"
flake8-bandit = "^2.1.2"
flake8-bugbear = "^21.4.3"
flake8-import-order = "^0.18.1"
black = "^21.6b0"
flake8-black = "^0.2.1"
pytest-cov = "^2.12.1"
coverage = "^5.5"
mypy = "^0.910"
typeguard = "^2.12.1"
pytest-mock = "^3.6.1"
requests-mock = "^1.9.3"
types-requests = "^2.25.0"
jupyterlab = "^3.0.16"
python-dotenv = "^0.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["src"]
omit = ["src/friend_network.py"] # TODO: Finish testing the file

[tool.coverage.report]
show_missing = true
fail_under = 100
